fix bug when try to delete item image

+++ src/main/java/dev/thiagooliveira/tablesplit/application/menu/UpdateItem.java
@@ -11,17 +11,11 @@ public class UpdateItem {
 
   private final GetItem getItem;
   private final ImageStorage imageStorage;
-  private final ItemImageRepository itemImageRepository;
   private final ItemRepository itemRepository;
 
-  public UpdateItem(
-      GetItem getItem,
-      ImageStorage imageStorage,
-      ItemImageRepository itemImageRepository,
-      ItemRepository itemRepository) {
+  public UpdateItem(GetItem getItem, ImageStorage imageStorage, ItemRepository itemRepository) {
     this.getItem = getItem;
     this.imageStorage = imageStorage;
-    this.itemImageRepository = itemImageRepository;
     this.itemRepository = itemRepository;
   }
 
@@ -31,32 +25,35 @@ public class UpdateItem {
     item.setDescription(command.description());
     item.setName(command.name());
     item.setCategoryId(command.categoryId());
-    this.itemRepository.save(item);
+
     var imageIdsToKeep =
         command.imageIdsToKeep() == null ? new ArrayList<>() : command.imageIdsToKeep();
-    item.getImages().stream()
-        .filter(img -> !imageIdsToKeep.contains(img.getId()))
-        .forEach(
+
+    item.getImages()
+        .removeIf(
             img -> {
-              var result = imageStorage.deleteItem(restaurantId, itemId, img.getId());
-              itemImageRepository.delete(img.getId());
+              if (!imageIdsToKeep.contains(img.getId())) {
+                imageStorage.deleteItem(restaurantId, itemId, img.getId());
+                return true;
+              }
+              return false;
             });
 
     if (command.images() != null && command.images().newImages() != null) {
       for (ImageData i : command.images().newImages()) {
         var image = new ItemImage();
         image.setId(UUID.randomUUID());
-        // For now, keep the simple logic, but ensure it doesn't break if already has
-        // main
+
         boolean hasMain =
             item.getImages().stream()
-                .anyMatch(img -> command.imageIdsToKeep().contains(img.getId()) && img.isMain());
+                .anyMatch(img -> imageIdsToKeep.contains(img.getId()) && img.isMain());
         image.setMain(!hasMain && command.images().newImages().indexOf(i) == 0);
         image.setItemId(item.getId());
         image.setName(
             imageStorage.uploadItem(i, item.getRestaurantId(), item.getId(), image.getId()));
-        this.itemImageRepository.save(image);
+        item.getImages().add(image);
       }
     }
+    this.itemRepository.save(item);
   }
 }
